+++ __test__/unit/core/modelCache.test.ts
@@ -1,7 +1,7 @@
+import ModelCache from '../../../src/core/caching/ModelCache';
 import { IdentityExecutor } from '../../../src/core/executors/IdentityExecutor';
 import { PropertiesExecutor } from '../../../src/core/executors/PropertiesExecutor';
 import { SubscriptionExecutor } from '../../../src/core/executors/SubscriptionExecutor';
-import ModelCache from '../../../src/core/caching/ModelCache';
 import { OSModel } from '../../../src/core/modelRepo/OSModel';
 import { ModelName } from '../../../src/core/models/SupportedModels';
 import { TestEnvironment } from '../../support/environment/TestEnvironment';
